refactor.User 가독성 향상을 위해 에러 핸들링 메소드, cors 설정 분리
