+++ be/src/config/corsConfig.ts
@@ -1,0 +1,7 @@
+import { CorsOptions } from '@nestjs/common/interfaces/external/cors-options.interface';
+
+export const corsOptions: CorsOptions = {
+  origin: 'http://localhost:3000',
+  methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
+  allowedHeaders: 'Content-Type, Accept',
+};
