+++ be/src/main.ts
@@ -1,16 +1,12 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { GlobalExceptionFilter } from './common/filters/GlobalExceptionFilter';
+import { corsOptions } from "./config/corsConfig";
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
 
-  // CORS 설정
-  app.enableCors({
-    origin: 'http://localhost:3000',
-    methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
-    allowedHeaders: 'Content-Type, Accept',
-  });
+  app.enableCors(corsOptions);
   app.useGlobalFilters(new GlobalExceptionFilter());
 
   await app.listen(3001);
